feat: move the core code to a separate module for reducing startup time
