+++ ftplugin/tex.lua
@@ -1,4 +1,4 @@
 local bufnr = vim.api.nvim_get_current_buf()
 vim.schedule(function()
-  require("math-conceal").set(nil, bufnr)
+  require("math-conceal.nvim").set(nil, bufnr)
 end)
